# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  agent.IndicatorView:
    properties:
      computedAt:
        example: "2026-09-17 15:30:00"
        type: string
      indicators: { }
      market:
        example: CN
        type: string
      period:
        example: daily
        type: string
      source:
        type: string
      symbol:
        example: 600519.SH
        type: string
      tradeDate:
        example: "2026-09-17"
        type: string
    type: object
...
  response.Envelope:
    properties:
      code:
        type: integer
      data: { }
      message:
        type: string
    type: object
...
info:
  contact: { }
  description: |-
    多智能体股票分析平台 HTTP 接口。

    所有响应共用同一个信封 {"code":0,"message":"ok","data":{...}}，
    code 为 0 表示成功，非 0 时 data 恒为 null，具体取值见 response 包。

    健康检查 GET /healthz 不在 /api/v1 之下，因此未收录于本文档。

    免责声明：本平台仅用于学习与研究，不构成投资建议。分析结论由大模型生成，
    可能包含事实性错误。过往表现不代表未来收益，投资有风险，可能损失本金。
  title: Trading Agents API
  version: "1.0"
paths:
...
  /agents/indicators/{code}:
    get:
      parameters:
        - description: 标的代码
          example: "600519"
          in: path
          name: code
          required: true
          type: string
        - description: 市场：CN/HK/US，留空按代码推断
          in: query
          name: market
          type: string
        - description: K 线周期：daily/weekly/monthly，默认 daily
          in: query
          name: period
          type: string
        - description: 交易日 YYYY-MM-DD，默认取当天；当天无行情则退到最近一个已算过的交易日
          in: query
          name: date
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/agent.IndicatorView'
                type: object
        "400":
          description: 代码、周期或交易日不合法
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 该标的尚无已落库的指标快照
          schema:
            $ref: '#/definitions/response.Envelope'
        "503":
          description: 技术指标仓储未就绪
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: 技术指标快照
      tags:
        - 智能体
...
securityDefinitions:
  BearerAuth:
    description: 填入 "Bearer {accessToken}"，令牌由 POST /auth/login 获取。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...